Dedupe Codex TOML line scanner and include path rewrite in CLI tsconfig
